move command frame history up a level
