+++ izzet/systems/clientsystems/clientsystems.go
@@ -6,7 +6,6 @@ import (
 	"github.com/kkevinchou/izzet/izzet/entities"
 	"github.com/kkevinchou/izzet/izzet/modellibrary"
 	"github.com/kkevinchou/izzet/izzet/network"
-	"github.com/kkevinchou/izzet/izzet/systems/clientsystems/commandframe"
 	"github.com/kkevinchou/kitolib/metrics"
 )
 
@@ -19,7 +18,7 @@ type App interface {
 	GetPlayerConnection() net.Conn
 	GetPlayerEntity() *entities.Entity
 	GetPlayerCamera() *entities.Entity
-	GetCommandFrameHistory() *commandframe.CommandFrameHistory
+	GetCommandFrameHistory() *CommandFrameHistory
 	MetricsRegistry() *metrics.MetricsRegistry
 	Client() network.IzzetClient
 }
